pyproject.toml
[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ['py37']

[tool.poetry]
name = "movee"
version = "0.9.2"
description = "🎦 movee: Script asciinema movies 🎦"
authors = ["Tom Ritchford <tom@swirly.com>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.8"
safer = "^4.4.1"
termtosvg = "^1.1.0"
pyyaml = "^6.0"
impall = "^1.1.1"
tdir = "^1.4.1"

[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
flake8 = "^5.0.4"
pytest = "^7.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
